==> Notebooks/0_city_data_web_scraping.ipynb <==
from bs4 import BeautifulSoup
import requests
import pandas as pd

url = "https://en.wikipedia.org/wiki/Berlin"

berlin = requests.get(url)
berlin.status_code

berlin_soup = BeautifulSoup(berlin.content, "html.parser")

name1 = berlin_soup.select(".infobox-full-data")[0].get_text()
name1

country1 = berlin_soup.select(".infobox-data a")[0].get_text()
country1

country_alt = berlin_soup.select(".infobox-data a")[0]["title"]
country_alt

lat1 = berlin_soup.select(".latitude")[0].get_text()
lat1

long1 = berlin_soup.select(".longitude")[0].get_text()
long1

elev1 = berlin_soup.select(".infobox-data")[5].get_text()
elev1

city_pop = berlin_soup.select(".infobox-data")[6].get_text()
city_pop

urban_pop1 = berlin_soup.select(".infobox-data")[7].get_text()
urban_pop1

metro_pop1 = berlin_soup.select(".infobox-data")[8].get_text()
metro_pop1

url = "https://en.wikipedia.org/wiki/Amsterdam"

amsterdam = requests.get(url)
amsterdam.status_code

amsterdam_soup = BeautifulSoup(amsterdam.content, "html.parser")

name2 = amsterdam_soup.select(".fn.org")[0].get_text()
name2

country2 = amsterdam_soup.select(".infobox-data a")[0].get_text()
country2

lat2 = amsterdam_soup.select(".latitude")[0].get_text()
lat2

long2 = amsterdam_soup.select(".longitude")[0].get_text()
long2

elev2 = amsterdam_soup.select(".infobox-data")[11].get_text()
elev2

muni_pop = amsterdam_soup.select(".infobox-data")[12].get_text()
muni_pop

urban_pop2 = amsterdam_soup.select(".infobox-data")[14].get_text()
urban_pop2

metro_pop2 = amsterdam_soup.select(".infobox-data")[15].get_text()
metro_pop2
